+++ infra/modules/worklytics-connector-specs/main.tf
@@ -3,6 +3,17 @@
 # more interoperability than in a .tf file
 
 
+# initial deployment time; in effect, timestamp of first `terraform apply`; will persist into the
+# terraform state.
+# we derive various API example calls from this; if we used `timestamp()` method, the example TODOs
+# and scripts would show changes on every apply
+# see https://registry.terraform.io/providers/hashicorp/time/latest/docs/resources/static
+# q: possibly better to declare this at root, and pass in as a variable? would reduce noise
+#    (one resource, instead of 3 because 'worklytics-connector-specs' is reference 3 times)
+resource "time_static" "deployment" {
+
+}
+
 locals {
 
   google_workspace_example_user  = coalesce(var.google_workspace_example_user, "REPLACE_WITH_EXAMPLE_USER@YOUR_COMPANY.COM")
@@ -82,7 +93,7 @@ locals {
       ],
       environment_variables : {},
       example_api_calls : [
-        "/gmail/v1/users/me/messages?maxResults=10",
+        "/gmail/v1/users/me/messages?maxResults=5&labelIds=SENT",
         "/gmail/v1/users/me/messages/{MESSAGE_ID}?format=metadata"
       ]
       example_api_calls_user_to_impersonate : local.google_workspace_example_user
@@ -218,7 +229,7 @@ locals {
       example_api_calls : [
         "/v1.0/users",
         "/v1.0/users/${var.example_msft_user_guid}/events",
-        "/v1.0/users/${var.example_msft_user_guid}/calendarView?startDateTime=2022-10-01T00:00:00Z&endDateTime=${timestamp()}",
+        "/v1.0/users/${var.example_msft_user_guid}/calendarView?startDateTime=2022-10-01T00:00:00Z&endDateTime=${time_static.deployment.id}",
         "/v1.0/users/${var.example_msft_user_guid}/mailboxSettings",
         "/v1.0/groups",
         "/v1.0/groups/{group-id}/members"
@@ -285,8 +296,8 @@ locals {
         "/beta/chats/${var.msft_teams_example_chat_guid}/messages",
         "/beta/communications/calls/${var.msft_teams_example_call_guid}",
         "/beta/communications/callRecords/${var.msft_teams_example_call_record_guid}",
-        "/beta/communications/callRecords/getDirectRoutingCalls(fromDateTime=${urlencode(timeadd(timestamp(), "-2160h"))},toDateTime=${urlencode(timestamp())})",
-        "/beta/communications/callRecords/getPstnCalls(fromDateTime=${urlencode(timeadd(timestamp(), "-2160h"))},toDateTime=${urlencode(timestamp())})",
+        "/beta/communications/callRecords/getDirectRoutingCalls(fromDateTime=${urlencode(timeadd(time_static.deployment.id, "-2160h"))},toDateTime=${urlencode(time_static.deployment.id)})",
+        "/beta/communications/callRecords/getPstnCalls(fromDateTime=${urlencode(timeadd(time_static.deployment.id, "-2160h"))},toDateTime=${urlencode(time_static.deployment.id)})",
         "/beta/users/${var.example_msft_user_guid}/onlineMeetings",
 
         "/v1.0/teams",
@@ -297,8 +308,8 @@ locals {
         "/v1.0/chats/${var.msft_teams_example_chat_guid}/messages",
         "/v1.0/communications/calls/${var.msft_teams_example_call_guid}",
         "/v1.0/communications/callRecords/${var.msft_teams_example_call_record_guid}",
-        "/v1.0/communications/callRecords/getDirectRoutingCalls(fromDateTime=${urlencode(timeadd(timestamp(), "-2160h"))},toDateTime=${urlencode(timestamp())})",
-        "/v1.0/communications/callRecords/getPstnCalls(fromDateTime=${urlencode(timeadd(timestamp(), "-2160h"))},toDateTime=${urlencode(timestamp())})",
+        "/v1.0/communications/callRecords/getDirectRoutingCalls(fromDateTime=${urlencode(timeadd(time_static.deployment.id, "-2160h"))},toDateTime=${urlencode(time_static.deployment.id)})",
+        "/v1.0/communications/callRecords/getPstnCalls(fromDateTime=${urlencode(timeadd(time_static.deployment.id, "-2160h"))},toDateTime=${urlencode(time_static.deployment.id)})",
         "/v1.0/users/${var.example_msft_user_guid}/onlineMeetings"
       ]
       external_token_todo : <<EOT
@@ -737,12 +748,12 @@ EOT
       example_api_calls : [
         "/services/data/v57.0/sobjects/Account/describe",
         "/services/data/v57.0/sobjects/ActivityHistory/describe",
-        "/services/data/v57.0/sobjects/Account/updated?start=${urlencode(timeadd(timestamp(), "-48h"))}&end=${urlencode(timestamp())}",
+        "/services/data/v57.0/sobjects/Account/updated?start=${urlencode(timeadd(time_static.deployment.id, "-48h"))}&end=${urlencode(time_static.deployment.id)}",
         "/services/data/v57.0/composite/sobjects/User?ids=${local.salesforce_example_account_id}&fields=Alias,AccountId,ContactId,CreatedDate,CreatedById,Email,EmailEncodingKey,Id,IsActive,LastLoginDate,LastModifiedDate,ManagerId,Name,TimeZoneSidKey,Username,UserRoleId,UserType",
         "/services/data/v57.0/composite/sobjects/Account?ids=${local.salesforce_example_account_id}&fields=Id,AnnualRevenue,CreatedDate,CreatedById,IsDeleted,LastActivityDate,LastModifiedDate,LastModifiedById,NumberOfEmployees,OwnerId,ParentId,Rating,Sic,Type",
         "/services/data/v57.0/query?q=SELECT%20%28SELECT%20AccountId%2CActivityDate%2CActivityDateTime%2CActivitySubtype%2CActivityType%2CCallDurationInSeconds%2CCallType%2CCreatedDate%2CCreatedById%2CDurationInMinutes%2CEndDateTime%2CId%2CIsAllDayEvent%2CIsDeleted%2CIsHighPriority%2CIsTask%2CLastModifiedDate%2CLastModifiedById%2COwnerId%2CPriority%2CStartDateTime%2CStatus%2CWhatId%2CWhoId%20FROM%20ActivityHistories%20ORDER%20BY%20LastModifiedDate%20DESC%20NULLS%20LAST%29%20FROM%20Account%20where%20id%3D%27${local.salesforce_example_account_id}%27",
-        "/services/data/v57.0/query?q=SELECT+Alias,AccountId,ContactId,CreatedDate,CreatedById,Email,EmailEncodingKey,Id,IsActive,LastLoginDate,LastModifiedDate,ManagerId,Name,TimeZoneSidKey,Username,UserRoleId,UserType+FROM+User+WHERE+LastModifiedDate+%3E%3D+${urlencode(timeadd(timestamp(), "-72h"))}+AND+LastModifiedDate+%3C+${urlencode(timestamp())}+ORDER+BY+LastModifiedDate+DESC+NULLS+LAST",
-        "/services/data/v57.0/query?q=SELECT+Id,AnnualRevenue,CreatedDate,CreatedById,IsDeleted,LastActivityDate,LastModifiedDate,LastModifiedById,NumberOfEmployees,OwnerId,ParentId,Rating,Sic,Type+FROM+Account+WHERE+LastModifiedDate+%3E%3D+${urlencode(timeadd(timestamp(), "-72h"))}+AND+LastModifiedDate+%3C+${urlencode(timestamp())}+ORDER+BY+LastModifiedDate+DESC+NULLS+LAST"
+        "/services/data/v57.0/query?q=SELECT+Alias,AccountId,ContactId,CreatedDate,CreatedById,Email,EmailEncodingKey,Id,IsActive,LastLoginDate,LastModifiedDate,ManagerId,Name,TimeZoneSidKey,Username,UserRoleId,UserType+FROM+User+WHERE+LastModifiedDate+%3E%3D+${urlencode(timeadd(time_static.deployment.id, "-72h"))}+AND+LastModifiedDate+%3C+${urlencode(time_static.deployment.id)}+ORDER+BY+LastModifiedDate+DESC+NULLS+LAST",
+        "/services/data/v57.0/query?q=SELECT+Id,AnnualRevenue,CreatedDate,CreatedById,IsDeleted,LastActivityDate,LastModifiedDate,LastModifiedById,NumberOfEmployees,OwnerId,ParentId,Rating,Sic,Type+FROM+Account+WHERE+LastModifiedDate+%3E%3D+${urlencode(timeadd(time_static.deployment.id, "-72h"))}+AND+LastModifiedDate+%3C+${urlencode(time_static.deployment.id)}+ORDER+BY+LastModifiedDate+DESC+NULLS+LAST"
       ]
       external_token_todo : <<EOT
   Before running the example, you have to populate the following variables in terraform:
@@ -1143,7 +1154,7 @@ EOT
       ],
       external_token_todo : <<EOT
 ## Prerequisites
-Jira OAuth 2.0 (3LO) through Psoxy requires a Jira Cloud account with following classical scopes:
+[Jira OAuth 2.0 (3LO)](https://developer.atlassian.com/cloud/jira/platform/oauth-2-3lo-apps/) through Psoxy requires a Jira Cloud account with following classical scopes:
 
 - read:jira-user: for getting generic user information
 - read:jira-work: for getting information about issues, comments, etc
@@ -1154,21 +1165,23 @@ And following granular scopes:
 - read:avatar:jira: for retrieving group members
 
 ## Setup Instructions
-  1. Go to https://developer.atlassian.com/console/myapps/ and click on "Create"
+  1. Go to https://developer.atlassian.com/console/myapps/ and click on "Create" and choose "OAuth 2.0 Integration"
 
-  2. Then click "Authorize" and "Add", adding `http://localhost` as callback URI. It can be any URL
-     that matches the settings.
+  2. Then click "Authorization" and "Add" on `OAuth 2.0 (3L0)`, adding `http://localhost` as callback URI. It can be any URL
+     that matches the URL format and it is required to be populated, but the proxy instance workflow will not use it.
 
-  3. Now navigate to "Permissions" and click on "Add" for Jira. Once added, click on "Configure".
-     Add following scopes as part of "Classic Scopes":
+  3. Now navigate to "Permissions" and click on "Add" for `Jira API`. Once added, click on "Configure".
+     Add following scopes as part of "Classic Scopes", first clicking on `Edit Scopes` and then selecting them:
        - `read:jira-user`
        - `read:jira-work`
      And these from "Granular Scopes":
        - `read:group:jira`
        - `read:avatar:jira`
        - `read:user:jira`
-     Then repeat the same but for "User Identity API", adding the following scope:
+     Then go back to "Permissions" and click on "Add" for `User Identity API`, only selecting following scopes:
        - `read:account`
+
+     After adding all the scopes, you should have 1 permission for `User Identity API` and 5 for `Jira API`.
 
   4. Once Configured, go to "Settings" and copy the "Client Id" and "Secret". You will use these to
      obtain an OAuth `refresh_token`.
@@ -1178,6 +1191,11 @@ And following granular scopes:
 
    `https://auth.atlassian.com/authorize?audience=api.atlassian.com&client_id=<CLIENT ID>&scope=offline_access%20read:group:jira%20read:avatar:jira%20read:user:jira%20read:account%20read:jira-user%20read:jira-work&redirect_uri=http://localhost&state=YOUR_USER_BOUND_VALUE&response_type=code&prompt=consent`
 
+     NOTES:
+     - That URL can be obtained from "Authorization" and clicking on `Configure` for  "OAuth 2.0 (3LO)" page.
+     - If after pasting the URL you see an error page (like an issue with identity provider) please check:
+       1 - The `clientId` is correct
+       2 - The `redirect_uri` is the same as the one configured in the "OAuth 2.0 (3LO)" page
   6. Choose a site in your Jira workspace to allow access for this application and click "Accept".
      As the callback does not exist, you will see an error. But in the URL of your browser you will see
      something like this as URL:
@@ -1212,7 +1230,7 @@ And following granular scopes:
      - `PSOXY_JIRA_CLOUD_CLIENT_ID` with `Client Id` value.
      - `PSOXY_JIRA_CLOUD_CLIENT_SECRET` with `Client Secret` value.
 
- 10. Optional, obtain the "Cloud ID" of your Jira instance. Use the following command, with the
+ 10. Obtain the "Cloud ID" of your Jira instance. Use the following command, with the
     `access_token` obtained in the previous step in place of `<ACCESS_TOKEN>` below:
 
    `curl --header 'Authorization: Bearer <ACCESS_TOKEN>' --url 'https://api.atlassian.com/oauth/token/accessible-resources'`
@@ -1233,7 +1251,7 @@ And following granular scopes:
 
   Add the `id` value from that JSON response as the value of the `jira_cloud_id` variable in the
   `terraform.tfvars` file of your Terraform configuration. This will generate all the test URLs with
-  a proper value.
+  a proper value and it will populate the right value for setting up the configuration.
 EOT
     }
   }