+++ tools/js-reference/index.js
@@ -1,0 +1,8 @@
+
+const { hashWithSalt } = require("./lib/lib.js");
+
+// Example usage
+const input = process.argv[1];
+const salt = process.argv[2];
+
+console.log(hashWithSalt(input, {salt: salt}));